Split ModelingSidebar out into own component
